+++ src/frontend/js/components/CourseProductCourseRuns/EnrolledCourseRun.tsx
@@ -1,8 +1,10 @@
 import { defineMessages, FormattedMessage } from 'react-intl';
+import { useMemo } from 'react';
 import { Spinner } from 'components/Spinner';
 import { useEnrollments } from 'hooks/useEnrollments';
 import type * as Joanie from 'types/Joanie';
 import useDateFormat from 'utils/useDateFormat';
+import useDateRelative from '../../utils/useDateRelative';
 import CourseRunSection, { messages as sectionMessages } from './CourseRunSection';
 
 const messages = defineMessages({
@@ -11,15 +13,25 @@ const messages = defineMessages({
     description: 'CTA displayed when user is enrolled to the course run.',
     id: 'components.EnrolledCourseRun.goToCourse',
   },
-  unroll: {
-    defaultMessage: 'Unroll',
-    description: 'Button label to unroll.',
-    id: 'components.EnrolledCourseRun.unroll',
+  isEnroll: {
+    defaultMessage: 'You are enrolled',
+    description: 'Message displayed when user is enrolled but the course run is not started',
+    id: 'components.EnrolledCourseRun.isEnroll',
   },
-  unrolling: {
-    defaultMessage: 'Unrolling...',
-    description: 'Accessible label displayed when user is being unrolled.',
-    id: 'components.EnrolledCourseRun.unrolling',
+  courseRunStartIn: {
+    defaultMessage: 'The course starts {relativeStartDate}',
+    description: 'Error displayed when user is enrolled but the course run is not started',
+    id: 'components.EnrolledCourseRun.courseRunStartIn',
+  },
+  unenroll: {
+    defaultMessage: 'Unenroll',
+    description: 'Button label to unenroll.',
+    id: 'components.EnrolledCourseRun.unenroll',
+  },
+  unenrolling: {
+    defaultMessage: 'Unenrolling...',
+    description: 'Accessible label displayed when user is being unenrolled.',
+    id: 'components.EnrolledCourseRun.unenrolling',
   },
 });
 
@@ -30,8 +42,15 @@ interface Props {
 const EnrolledCourseRun = ({ enrollment }: Props) => {
   const formatDate = useDateFormat();
   const { methods, states } = useEnrollments();
+  const relativeStartDate = useDateRelative(new Date(enrollment.course_run.start));
 
-  const unroll = () => {
+  const isStarded = useMemo(() => {
+    const startDateTime = new Date(enrollment.course_run.start);
+    const today = new Date();
+    return startDateTime <= today;
+  }, [enrollment]);
+
+  const unenroll = () => {
     methods.update({
       course_run: enrollment.course_run.id,
       is_active: false,
@@ -71,22 +90,37 @@ const EnrolledCourseRun = ({ enrollment }: Props) => {
           </em>
         </div>
         {states.error && <p className="course-runs-list__errors-feedback ">{states.error}</p>}
+
+        {!isStarded && (
+          <p className="course-runs-list__errors-feedback ">
+            <FormattedMessage {...messages.courseRunStartIn} values={{ relativeStartDate }} />
+          </p>
+        )}
         <div className="course-runs-item">
-          <a
-            href={enrollment.course_run.resource_link}
-            className="course-runs-item__cta button--primary button--pill button--tiny"
-          >
-            <FormattedMessage {...messages.goToCourse} />
-          </a>
-          <button className="button--tiny" onClick={unroll}>
+          {isStarded && (
+            <a
+              href={enrollment.course_run.resource_link}
+              className="course-runs-item__cta button--primary button--pill button--tiny"
+            >
+              <FormattedMessage {...messages.goToCourse} />
+            </a>
+          )}
+
+          {!isStarded && (
+            <p className="course-runs-item__enrolled button--primary button--pill button--tiny disabled">
+              <FormattedMessage {...messages.isEnroll} />
+            </p>
+          )}
+
+          <button className="button--tiny" onClick={unenroll}>
             {states.updating ? (
               <Spinner aria-labelledby={`unrolling-${enrollment.id}`}>
                 <span id={`unrolling-${enrollment.id}`}>
-                  <FormattedMessage {...messages.unrolling} />
+                  <FormattedMessage {...messages.unenrolling} />
                 </span>
               </Spinner>
             ) : (
-              <FormattedMessage {...messages.unroll} />
+              <FormattedMessage {...messages.unenroll} />
             )}
           </button>
         </div>